pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dw6"
version = "0.1.0"
dependencies = [
    "fastapi",
    "uvicorn",
    "python-dotenv",
    "GitPython",
    "httpx[socks]",
    "httpx",
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-cov",
    "pytest-mock",
]

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
dw6 = "dw6.cli:main"
